Updated vcr spec tests
